pyproject.toml
[project]
name = "suwalski-assistant"
version = "0.1.0"
description = "Core hub of a AI-powered system designed for seamless interaction, tool orchestration, and multi-agent workflows."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "fastmcp",
    "litellm",
    "discord.py",
    "python-dotenv",
    "Pillow"
]

[project.scripts]
start = "suwalski_assistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/suwalski_assistant"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

# Custom scripts configuration (conceptual, assuming usage with a task runner or uv)
# creating-deployable-bundle = "scripts/bundle.sh" # Placeholder for bundle script
